+++ figure.cpp
@@ -2,7 +2,7 @@
 
 #include "include/figure.h"
 
-void Figure::Move(int px, int py) {
+void Figure::Move(double px, double py) {
     pos.x = px;
     pos.y = py;
 }
